Adds heart.system namespace functions, corresponding to love.system
This changeset implements all but love.system.getProcessorCount(), because that
particular function can't be done cleanly in JS without pulling in an external
library https://github.com/oftn/core-estimator . Whether pulling in that
library to enable one, likely rarely-used LÖVE function is worth it is up for
debate.

All the functions work to some extent in all the browsers, but due to security
restrictions a lot of it does not work as nicely as one would wish. Also, the
HTML5 Battery Status API is not available in browsers other than Firefox yet,
and there is no other way to access that information; fortunately that API will
soon be implemented in Chrome (it's already in WebKit itself).

--HG--
rename : examples/basic.html => examples/system.html